move sources to match maven structure
src/main/java  src/test/java
