+++ .gitignore
@@ -7,3 +7,5 @@ doc/
 *.gch
 *.sublime-workspace
 .README.md.html
+/target/
+.settings/
