+++ src/main/java/org/magnos/trie/PerfectHashMap.java
