+++ src/main/java/org/magnos/trie/TrieMatch.java
