+++ src/main/java/org/magnos/trie/TrieNode.java
