+++ src/main/java/org/magnos/trie/TrieSequencer.java
