+++ src/main/java/org/magnos/trie/TrieSequencerByteArray.java
