+++ src/main/java/org/magnos/trie/TrieSequencerByteBuffer.java
