+++ src/main/java/org/magnos/trie/TrieSequencerCharArray.java
