+++ src/main/java/org/magnos/trie/TrieSequencerCharArrayCaseInsensitive.java
