+++ src/main/java/org/magnos/trie/TrieSequencerCharSequence.java
