+++ src/main/java/org/magnos/trie/TrieSequencerCharSequenceCaseInsensitive.java
