+++ src/main/java/org/magnos/trie/TrieSequencerIntArray.java
