+++ src/main/java/org/magnos/trie/TrieSequencerLongArray.java
