+++ src/main/java/org/magnos/trie/TrieSequencerShortArray.java
