+++ src/main/java/org/magnos/trie/Tries.java
