+++ src/test/java/org/magnos/trie/TestAutoComplete.java
