+++ src/test/java/org/magnos/trie/TestPerfectHashMap.java
