+++ src/test/java/org/magnos/trie/TestTrie.java
