+++ src/test/java/org/magnos/trie/TestTrieSequencerCharArray.java
