+++ src/test/java/org/magnos/trie/TestTrieSequencerCharArrayCaseInsensitive.java
