+++ src/test/java/org/magnos/trie/TestTrieSequencerCharSequence.java
