+++ src/test/java/org/magnos/trie/TestTrieSequencerCharSequenceCaseInsensitive.java
