+++ src/test/java/org/magnos/trie/dictionary.txt
